ExamLens/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
django-cors-headers>=4.3
whitenoise>=6.7
dj-database-url>=2.1
psycopg2-binary>=2.9
